+++ example/android/app/src/main/kotlin/com/example/iconslauncher/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.default_example
+package com.example.iconslauncher
 
 import io.flutter.embedding.android.FlutterActivity