+++ lib/xml_templates.dart
@@ -14,6 +14,22 @@ const String icLauncherDrawableBackgroundXml = '''
 </adaptive-icon>
 ''';
 
+const String icLauncherMipMapXml = '''
+<?xml version="1.0" encoding="utf-8"?>
+<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
+  <background android:drawable="@mipmap/ic_launcher_background"/>
+  <foreground android:drawable="@mipmap/ic_launcher_foreground"/>
+</adaptive-icon>
+''';
+
+const String icLauncherRoundMipMapXml = '''
+<?xml version="1.0" encoding="utf-8"?>
+<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
+  <background android:drawable="@mipmap/ic_launcher_background"/>
+  <foreground android:drawable="@mipmap/ic_launcher_foreground"/>
+</adaptive-icon>
+''';
+
 const String colorsXml = '''
 <?xml version="1.0" encoding="utf-8"?>
 <resources>