+++ test/main_test.dart
@@ -1,20 +1,30 @@
-import 'package:args/args.dart';
-import 'package:icons_launcher/main.dart' show defaultConfigFile;
+import 'package:icons_launcher/constants.dart';
 import 'package:icons_launcher/main.dart' as main_dart;
 import 'package:icons_launcher/src/android.dart' as android;
 import 'package:icons_launcher/src/ios.dart' as ios;
-import 'package:path/path.dart';
 import 'package:test/test.dart';
 import 'package:universal_io/io.dart';
+import 'package:yaml/yaml.dart';
 
 // Unit tests for main.dart
 void main() {
+  test('Version checker', () {
+    // Read yaml config file
+    final File file = File('pubspec.yaml');
+    final String content = file.readAsStringSync();
+    final Map yaml = loadYaml(content);
+
+    // Read message constant
+    final message = introMessage();
+    expect(message, contains(yaml['version']));
+  });
+
   test('Android icon list is correct size', () {
     expect(android.androidIcons.length, 5);
   });
 
   test('Adaptive foreground icons icon list is correct size', () {
-    expect(android.adaptiveForegroundIcons.length, 6);
+    expect(android.adaptiveForegroundIcons.length, 5);
   });
 
   test('iOS icon list is correct size', () {
@@ -33,81 +43,6 @@ void main() {
     expect(config.length, isNotNull);
   });
 
-  group('config file from args', () {
-    // Create mini parser with only the wanted option, mocking the real one
-    final ArgParser parser = ArgParser()
-      ..addOption(main_dart.fileOption, abbr: 'f');
-    final String testDir =
-        join('.dart_tool', 'icons_launcher', 'test', 'config_file');
-
-    late String currentDirectory;
-    Future<void> setCurrentDirectory(String path) async {
-      path = join(testDir, path);
-      await Directory(path).create(recursive: true);
-      Directory.current = path;
-    }
-
-    setUp(() {
-      currentDirectory = Directory.current.path;
-    });
-    tearDown(() {
-      Directory.current = currentDirectory;
-    });
-    test('default', () async {
-      await setCurrentDirectory('default');
-      await File('icons_launcher.yaml').writeAsString('''
-flutter_icons:
-  android: true
-  ios: false
-''');
-      final ArgResults argResults = parser.parse(<String>[]);
-      final Map<String, dynamic>? config =
-          main_dart.loadConfigFileFromArgResults(argResults);
-      expect(config, isNotNull);
-      expect(config!['android'], true);
-    });
-    test('default_use_pubspec', () async {
-      await setCurrentDirectory('pubspec_only');
-      await File('pubspec.yaml').writeAsString('''
-flutter_icons:
-  android: true
-  ios: false
-''');
-      ArgResults argResults = parser.parse(<String>[]);
-      final Map<String, dynamic>? config =
-          main_dart.loadConfigFileFromArgResults(argResults);
-      expect(config, isNotNull);
-      expect(config!['ios'], false);
-
-      // fails if forcing default file
-      argResults = parser.parse(<String>['-f', defaultConfigFile]);
-      expect(main_dart.loadConfigFileFromArgResults(argResults), isNull);
-    });
-
-    test('custom', () async {
-      await setCurrentDirectory('custom');
-      await File('custom.yaml').writeAsString('''
-flutter_icons:
-  android: true
-  ios: true
-''');
-      // if no argument set, should fail
-      ArgResults argResults = parser.parse(<String>['-f', 'custom.yaml']);
-      final Map<String, dynamic>? config =
-          main_dart.loadConfigFileFromArgResults(argResults);
-      expect(config, isNotNull);
-      expect(config!['ios'], true);
-
-      // should fail if no argument
-      argResults = parser.parse(<String>[]);
-      expect(main_dart.loadConfigFileFromArgResults(argResults), isNull);
-
-      // or missing file
-      argResults = parser.parse(<String>['-f', 'missing_custom.yaml']);
-      expect(main_dart.loadConfigFileFromArgResults(argResults), isNull);
-    });
-  });
-
   test('Incorrect pubspec.yaml path throws correct error message', () async {
     const String incorrectPath = 'test/config/test_pubspec.yam';
     expect(() => main_dart.loadConfigFile(incorrectPath, null),
@@ -116,7 +51,7 @@ flutter_icons:
 
   test('image_path is in config', () {
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
-      'image_path': 'assets/images/icon-710x599.png',
+      'image_path': 'assets/ic_logo_border.png',
       'android': true,
       'ios': true
     };
@@ -128,24 +63,30 @@ flutter_icons:
     };
     expect(main_dart.isImagePathInConfig(flutterIconsConfigAndroid), false);
     final Map<String, dynamic> flutterIconsConfigBoth = <String, dynamic>{
-      'image_path_android': 'assets/images/icon-710x599.png',
-      'image_path_ios': 'assets/images/icon-710x599.png',
-      'image_path_macos': 'assets/images/icon-710x599.png',
-      'image_path_windows': 'assets/images/icon-710x599.png',
-      'adaptive_icon_background': 'assets/images/christmas-background.png',
-      'adaptive_icon_foreground': 'assets/images/icon-foreground-432x432.png',
-      'remove_alpha_ios': true,
-      'android': true,
+      'image_path_android': 'assets/ic_logo_border.png',
+      'image_path_ios': 'assets/ic_logo_rectangle.png',
+      'image_path_macos': 'assets/ic_logo_border.png',
+      'image_path_windows': 'assets/ic_logo_border.png',
+      'image_path_linux': 'assets/ic_logo_border.png',
+      'image_path_web': 'assets/ic_logo_border.png',
+      'adaptive_icon_background': 'assets/ic_background.png',
+      'adaptive_icon_foreground': 'assets/ic_foreground.png',
+      'adaptive_icon_round': 'assets/ic_logo_round.png',
+      'remove_alpha_ios': false,
+      'remove_alpha_macos': false,
       'ios': true,
-      'macos': true,
-      'windows': true
+      'android': true,
+      'macos': false,
+      'windows': false,
+      'linux': false,
+      'web': false
     };
     expect(main_dart.isImagePathInConfig(flutterIconsConfigBoth), true);
   });
 
   test('At least one platform is in config file', () {
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
-      'image_path': 'assets/images/icon-710x599.png',
+      'image_path': 'assets/ic_logo_border.png',
       'android': true,
       'ios': true
     };
@@ -154,14 +95,14 @@ flutter_icons:
 
   test('No platform specified in config', () {
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
-      'image_path': 'assets/images/icon-710x599.png'
+      'image_path': 'assets/ic_logo_border.png'
     };
     expect(main_dart.hasPlatformConfig(flutterIconsConfig), false);
   });
 
   test('No new Android icon needed - android: false', () {
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
-      'image_path': 'assets/images/icon-710x599.png',
+      'image_path': 'assets/ic_logo_border.png',
       'android': false,
       'ios': true
     };
@@ -170,7 +111,7 @@ flutter_icons:
 
   test('No new Android icon needed - no Android config', () {
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
-      'image_path': 'assets/images/icon-710x599.png',
+      'image_path': 'assets/ic_logo_border.png',
       'ios': true
     };
     expect(main_dart.isNeedingNewAndroidIcon(flutterIconsConfig), false);
@@ -178,7 +119,7 @@ flutter_icons:
 
   test('No new iOS icon needed - ios: false', () {
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
-      'image_path': 'assets/images/icon-710x599.png',
+      'image_path': 'assets/ic_logo_border.png',
       'android': true,
       'ios': false
     };
@@ -187,7 +128,7 @@ flutter_icons:
 
   test('No new iOS icon needed - no iOS config', () {
     final Map<String, dynamic> flutterIconsConfig = <String, dynamic>{
-      'image_path': 'assets/images/icon-710x599.png',
+      'image_path': 'assets/ic_logo_border.png',
       'android': true
     };
     expect(main_dart.isNeedingNewIOSIcon(flutterIconsConfig), false);